requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.3.1
python-craigslist==1.0.6
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
slackclient==1.3.1
soupsieve==1.9.1
SQLAlchemy==1.3.3
typed-ast==1.3.4
urllib3==1.24.2
websocket-client==0.54.0
wrapt==1.11.1
